Break up profiles router
