+++ src/proxy/http/profiles/mod.rs
@@ -1,0 +1,284 @@
+extern crate tower_discover;
+
+use futures::Stream;
+use http;
+use indexmap::IndexMap;
+// Import is used by WeightedIndex::sample.
+#[allow(unused_imports)]
+use rand::distributions::Distribution;
+use rand::distributions::WeightedIndex;
+use regex::Regex;
+use std::fmt;
+use std::hash::{Hash, Hasher};
+use std::iter::FromIterator;
+use std::ops::Deref;
+use std::sync::Arc;
+use std::time::Duration;
+
+use never::Never;
+
+use super::retry::Budget;
+use NameAddr;
+
+/// A stack module that produces a Service that routes requests through alternate
+/// middleware configurations
+///
+/// As the router's Stack is built, a destination is extracted from the stack's
+/// target and it is used to get route profiles from ` GetRoutes` implemetnation.
+///
+/// Each route uses a shared underlying concrete dst router.  The concrete dst
+/// router picks a concrete dst (NameAddr) from the profile's `dst_overrides` if
+/// they exist, or uses the router's target's addr if no `dst_overrides` exist.
+/// The concrete dst router uses the concrete dst as the target for the
+/// underlying stack.
+pub mod router;
+
+#[derive(Clone)]
+pub struct WeightedAddr {
+    pub addr: NameAddr,
+    pub weight: u32,
+}
+
+pub struct Routes {
+    pub routes: Vec<(RequestMatch, Route)>,
+    pub dst_overrides: Vec<WeightedAddr>,
+}
+
+/// Watches a destination's Routes.
+///
+/// The stream updates with all routes for the given destination. The stream
+/// never ends and cannot fail.
+pub trait GetRoutes {
+    type Stream: Stream<Item = Routes, Error = Never>;
+
+    fn get_routes(&self, dst: &NameAddr) -> Option<Self::Stream>;
+}
+
+/// Implemented by target types that may be combined with a Route.
+pub trait WithRoute {
+    type Output;
+
+    fn with_route(self, route: Route) -> Self::Output;
+}
+
+/// Implemented by target types that can have their `NameAddr` destination
+/// changed.
+pub trait WithAddr {
+    fn with_addr(self, addr: NameAddr) -> Self;
+}
+
+/// Implemented by target types that may have a `NameAddr` destination that
+/// can be discovered via `GetRoutes`.
+pub trait CanGetDestination {
+    fn get_destination(&self) -> Option<&NameAddr>;
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Eq, Hash)]
+pub struct Route {
+    labels: Labels,
+    response_classes: ResponseClasses,
+    retries: Option<Retries>,
+    timeout: Option<Duration>,
+}
+
+#[derive(Clone, Debug)]
+pub enum RequestMatch {
+    All(Vec<RequestMatch>),
+    Any(Vec<RequestMatch>),
+    Not(Box<RequestMatch>),
+    Path(Regex),
+    Method(http::Method),
+}
+
+#[derive(Clone, Debug)]
+pub struct ResponseClass {
+    is_failure: bool,
+    match_: ResponseMatch,
+}
+
+#[derive(Clone, Default)]
+pub struct ResponseClasses(Arc<Vec<ResponseClass>>);
+
+#[derive(Clone, Debug)]
+pub enum ResponseMatch {
+    All(Vec<ResponseMatch>),
+    Any(Vec<ResponseMatch>),
+    Not(Box<ResponseMatch>),
+    Status {
+        min: http::StatusCode,
+        max: http::StatusCode,
+    },
+}
+
+#[derive(Clone, Debug)]
+pub struct Retries {
+    budget: Arc<Budget>,
+}
+
+#[derive(Clone, Default)]
+struct Labels(Arc<IndexMap<String, String>>);
+
+// === impl Route ===
+
+impl Route {
+    pub fn new<I>(label_iter: I, response_classes: Vec<ResponseClass>) -> Self
+    where
+        I: Iterator<Item = (String, String)>,
+    {
+        let labels = {
+            let mut pairs = label_iter.collect::<Vec<_>>();
+            pairs.sort_by(|(k0, _), (k1, _)| k0.cmp(k1));
+            Labels(Arc::new(IndexMap::from_iter(pairs)))
+        };
+
+        Self {
+            labels,
+            response_classes: ResponseClasses(response_classes.into()),
+            retries: None,
+            timeout: None,
+        }
+    }
+
+    pub fn labels(&self) -> &Arc<IndexMap<String, String>> {
+        &self.labels.0
+    }
+
+    pub fn response_classes(&self) -> &ResponseClasses {
+        &self.response_classes
+    }
+
+    pub fn retries(&self) -> Option<&Retries> {
+        self.retries.as_ref()
+    }
+
+    pub fn timeout(&self) -> Option<Duration> {
+        self.timeout
+    }
+
+    pub fn set_retries(&mut self, budget: Arc<Budget>) {
+        self.retries = Some(Retries { budget });
+    }
+
+    pub fn set_timeout(&mut self, timeout: Duration) {
+        self.timeout = Some(timeout);
+    }
+}
+
+// === impl RequestMatch ===
+
+impl RequestMatch {
+    fn is_match<B>(&self, req: &http::Request<B>) -> bool {
+        match self {
+            RequestMatch::Method(ref method) => req.method() == *method,
+            RequestMatch::Path(ref re) => re.is_match(req.uri().path()),
+            RequestMatch::Not(ref m) => !m.is_match(req),
+            RequestMatch::All(ref ms) => ms.iter().all(|m| m.is_match(req)),
+            RequestMatch::Any(ref ms) => ms.iter().any(|m| m.is_match(req)),
+        }
+    }
+}
+
+// === impl ResponseClass ===
+
+impl ResponseClass {
+    pub fn new(is_failure: bool, match_: ResponseMatch) -> Self {
+        Self { is_failure, match_ }
+    }
+
+    pub fn is_failure(&self) -> bool {
+        self.is_failure
+    }
+
+    pub fn is_match<B>(&self, req: &http::Response<B>) -> bool {
+        self.match_.is_match(req)
+    }
+}
+
+// === impl ResponseClasses ===
+
+impl Deref for ResponseClasses {
+    type Target = [ResponseClass];
+
+    fn deref(&self) -> &Self::Target {
+        &*self.0
+    }
+}
+
+impl PartialEq for ResponseClasses {
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.0, &other.0)
+    }
+}
+
+impl Eq for ResponseClasses {}
+
+impl Hash for ResponseClasses {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_usize(Arc::as_ref(&self.0) as *const _ as usize);
+    }
+}
+
+impl fmt::Debug for ResponseClasses {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+// === impl ResponseMatch ===
+
+impl ResponseMatch {
+    fn is_match<B>(&self, req: &http::Response<B>) -> bool {
+        match self {
+            ResponseMatch::Status { ref min, ref max } => {
+                *min <= req.status() && req.status() <= *max
+            }
+            ResponseMatch::Not(ref m) => !m.is_match(req),
+            ResponseMatch::All(ref ms) => ms.iter().all(|m| m.is_match(req)),
+            ResponseMatch::Any(ref ms) => ms.iter().any(|m| m.is_match(req)),
+        }
+    }
+}
+
+// === impl Retries ===
+
+impl Retries {
+    pub fn budget(&self) -> &Arc<Budget> {
+        &self.budget
+    }
+}
+
+impl PartialEq for Retries {
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.budget, &other.budget)
+    }
+}
+
+impl Eq for Retries {}
+
+impl Hash for Retries {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_usize(Arc::as_ref(&self.budget) as *const _ as usize);
+    }
+}
+
+// === impl Labels ===
+
+impl PartialEq for Labels {
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.0, &other.0)
+    }
+}
+
+impl Eq for Labels {}
+
+impl Hash for Labels {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_usize(Arc::as_ref(&self.0) as *const _ as usize);
+    }
+}
+
+impl fmt::Debug for Labels {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
